+++ generators/app/index.js
@@ -1,7 +1,7 @@
 const Generator = require('yeoman-generator');
 const { gray, red, yellow } = require('chalk').default;
 const yosay = require('yosay');
-const { join } = require('path');
+const path = require('path');
 const validate = require('validate-npm-package-name');
 
 const SEQUELIZE_DIALECT = require('./utils/sequelize-dialect-enum');
@@ -12,6 +12,7 @@ module.exports = class extends Generator {
 
     this.answers = await this.prompt([
       {
+        store: true,
         type: 'input',
         name: 'shortname',
         message: `Package ${yellow('shortname')} [ex: ${gray('rest-api')}]`,
@@ -20,76 +21,174 @@ module.exports = class extends Generator {
           'Must be a valid package name',
       },
       {
+        store: true,
         type: 'input',
         name: 'name',
         message: `App ${yellow('(pretty) name')} [ex: ${gray('My REST API')}]`,
         validate: input => input.trim() !== '',
       },
       {
+        store: true,
         type: 'input',
         name: 'description',
         message: `App ${yellow('description')} [ex: ${gray('A REST API for cats and dogs')}]`,
         validate: input => input.trim() !== '',
       },
+      //
+      // ─── SEQUELIZE ───────────────────────────────────────────────────
+      //
       {
+        store: true,
         type: 'confirm',
         name: 'sequelize',
         message: `Use ${yellow.bold('Sequelize')} as ORM ?`,
         default: true,
       },
       {
+        store: true,
         type: 'list',
         name: 'sequelizeDialect',
-        message: `What is your database ${yellow.bold('dialect')} ?`,
+        message: `What is your ${yellow.bold('database dialect')} ?`,
         choices: Object.values(SEQUELIZE_DIALECT),
-        when: ({ sequelize }) => {
-          // Only run if user set Sequelize
-          return !!sequelize;
-        },
+        when: ({ sequelize }) => !!sequelize,
       },
+      //
+      // ─── REDIS ───────────────────────────────────────────────────────
+      //
       {
+        store: true,
+        type: 'confirm',
+        name: 'redis',
+        message: `Use ${yellow.bold('Redis')} for caching ?`,
+        default: true,
+      },
+      //
+      // ─── SOCKETIO ────────────────────────────────────────────────────
+      //
+      {
+        store: true,
+        type: 'confirm',
+        name: 'socketIo',
+        message: `Use ${yellow.bold('Socket.IO')} for real-time events ?`,
+        default: true,
+      },
+      //
+      // ─── I18N ────────────────────────────────────────────────────────
+      //
+      {
+        store: true,
+        type: 'confirm',
+        name: 'i18n',
+        message: `Use ${yellow.bold('i18next')} for internationalization ?`,
+        default: true,
+      },
+      //
+      // ─── WINSTON ─────────────────────────────────────────────────────
+      //
+      {
+        store: true,
         type: 'confirm',
         name: 'winston',
-        message: `Use ${yellow.bold('winston')} for logging ?`,
+        message: `Use ${yellow.bold('Winston')} for logging ?`,
         default: true,
       },
+      //
+      // ─── CELEBRATE ───────────────────────────────────────────────────
+      //
       {
+        store: true,
         type: 'confirm',
         name: 'celebrate',
-        message: `Use ${yellow.bold('celebrate')} and ${yellow.bold('Joi')} for object validation ?`,
+        message: `Use ${yellow.bold('celebrate')} and ${yellow.bold(
+          'Joi'
+        )} for object validation ?`,
         default: true,
       },
+      //
+      // ─── JWT ─────────────────────────────────────────────────────────
+      //
       {
+        store: true,
         type: 'confirm',
         name: 'jwt',
         message: `Use ${yellow.bold('JWT')} for user authentication ?`,
         default: true,
       },
+      //
+      // ─── NODEMAILER ──────────────────────────────────────────────────
+      //
       {
+        store: true,
         type: 'confirm',
         name: 'nodemailer',
-        message: `Use ${yellow.bold('nodemailer')} to send emails ?`,
+        message: `Use ${yellow.bold('Nodemailer')} to send emails ?`,
         default: true,
       },
+      //
+      // ─── CRON ────────────────────────────────────────────────────────
+      //
       {
+        store: true,
+        type: 'confirm',
+        name: 'cron',
+        message: `Use ${yellow.bold('node-schedule')} to handle cron jobs ?`,
+        default: true,
+      },
+      //
+      // ─── ESLINT ──────────────────────────────────────────────────────
+      //
+      {
+        store: true,
+        type: 'confirm',
+        name: 'eslint',
+        message: `Use ${yellow.bold('ESLint')} for code linting ?`,
+        default: true,
+      },
+      //
+      // ─── PRETTIER ────────────────────────────────────────────────────
+      //
+      {
+        store: true,
         type: 'confirm',
         name: 'prettier',
         message: `Use ${yellow.bold('Prettier')} for code formatting ?`,
         default: true,
       },
+      //
+      // ─── MOCHA ───────────────────────────────────────────────────────
+      //
       {
+        store: true,
         type: 'confirm',
         name: 'mocha',
         message: `Use ${yellow.bold('Mocha')} and ${yellow.bold('Chai')} for testing ?`,
         default: true,
       },
+      //
+      // ─── SENTRY ──────────────────────────────────────────────────────
+      //
       {
+        store: true,
         type: 'confirm',
-        name: 'docker',
-        message: `Generate a ${yellow('DockerFile')} ?`,
+        name: 'sentry',
+        message: `Use ${yellow.bold('Sentry')} for error tracking ?`,
         default: true,
       },
+      //
+      // ─── DOCKER ──────────────────────────────────────────────────────
+      //
       {
+        store: true,
+        type: 'confirm',
+        name: 'docker',
+        message: `Generate ${yellow('Docker')} related files ?`,
+        default: true,
+      },
+      //
+      // ─── OPENAPI ─────────────────────────────────────────────────────
+      //
+      {
+        store: true,
         type: 'confirm',
         name: 'openapi',
         message: `Generate an ${yellow.bold('OpenAPI')} documentation file ?`,
@@ -101,163 +200,384 @@ module.exports = class extends Generator {
   writing() {
     const { answers } = this;
     const { shortname } = answers;
-    const copy = this.fs.copy.bind(this.fs);
-    const copyTpl = this.fs.copyTpl.bind(this.fs);
     const src = this.templatePath.bind(this);
-    const dest = path => this.destinationPath.bind(this)(`${shortname}/${path}`);
+    const dest = to => this.destinationPath.bind(this)(`${shortname}/${to}`);
+    const copy = (from, data) =>
+      this.fs.copyTpl.bind(this.fs)(src(from), dest(from.replace('.ejs', '')), data || answers);
 
     this.packages = {
       dependencies: [
-        '@kazaar/express-error-handler@2',
+        '@kazaar/express-error-handler@3',
         'body-parser@1',
+        'compression@1',
         'convict@5',
+        'cors@2',
         'dotenv@8',
         'express@4',
         'express-promise-router@3',
-        'helmet@3',
+        'express-rate-limit@5',
+        'express-status-monitor@1',
+        'helmet@4',
         'http-errors@1',
         'moment@2',
+        'moment-timezone@0.5',
         'morgan@1',
+        'reflect-metadata@0',
+        'serve-favicon@2',
+        'typedi@0',
       ],
       devDependencies: [
-        'cross-env@5',
-        'eslint@5',
-        'eslint-config-airbnb-base@1',
-        'eslint-plugin-import@2',
-        'eslint-plugin-node@8',
-        'eslint-plugin-promise@4',
-        'husky@1',
-        'lint-staged@8',
-        'nodemon@1',
+        '@types/compression@1',
+        '@types/convict@4',
+        '@types/cors@2',
+        '@types/express@4',
+        '@types/express-rate-limit@5',
+        '@types/express-status-monitor@1',
+        '@types/http-errors@1',
+        '@types/morgan@1',
+        '@types/node@12',
+        '@types/serve-favicon@2',
+        'cpx@1',
+        'cross-env@7',
+        'husky@4',
+        'lint-staged@10',
         'npm-run-all@4',
+        'rimraf@3',
+        'ts-node@8',
+        'tsc-watch@4',
+        'typescript@4',
       ],
     };
 
-    const dialect = Object.values(SEQUELIZE_DIALECT).find(d => d.value === answers.sequelizeDialect);
+    const dialect = Object.values(SEQUELIZE_DIALECT).find(
+      d => d.value === answers.sequelizeDialect
+    );
 
-    const envName = shortname.toUpperCase().replace(/-/g, '_');
+    answers.auth = !!answers.sequelize && !!answers.jwt;
+    answers.description = JSON.stringify(answers.description);
+    answers.dialect = dialect;
 
-    copy(src('editorconfig'), dest('.editorconfig'));
-    copy(src('gitattributes'), dest('.gitattributes'));
-    copy(src('gitignore'), dest('.gitignore'));
-    copy(src('nodemon'), dest('nodemon.json'));
-    copy(src('huskyrc'), dest('.huskyrc.json'));
+    copy('.editorconfig.ejs');
+    copy('.gitattributes.ejs');
+    copy('src/dto/index.ts.ejs');
+    copy('src/config/index.ts.ejs');
+    copy('src/config/logger.ts.ejs');
+    copy('src/loaders/index.ts.ejs');
+    copy('src/middlewares/error-logger.ts.ejs');
+    copy('src/middlewares/index.ts.ejs');
+    copy('src/public/favicon.ico');
+    copy('src/routes/index.ts.ejs');
+    copy('src/services/index.ts.ejs');
+    copy('src/types/enums/index.ts.ejs');
+    copy('src/types/index.d.ts.ejs');
+    copy('src/types/index.ts.ejs');
+    copy('src/index.ts.ejs');
+    copy('src/server.ts.ejs');
+    copy('.env.ejs');
+    copy('.env.example.ejs');
+    copy('.gitignore.ejs');
+    copy('.huskyrc.json.ejs');
+    copy('ABOUT.md.ejs');
+    copy('README.md.ejs');
+    copy('package.json.ejs');
+    copy('tsconfig.json.ejs');
+    copy('tsconfig.build.json.ejs');
 
-    copy(src('src/api/controllers/empty'), dest('src/api/controllers/empty'));
-    copy(src('src/api/routes/empty'), dest('src/api/routes/empty'));
-    copy(src('src/services/empty'), dest('src/services/empty'));
-
-    copyTpl(src('eslintrc'), dest('.eslintrc.json'), answers);
-    copyTpl(src('README'), dest('README.md'), answers);
-    copyTpl(src('CONTRIBUTING'), dest('CONTRIBUTING.md'), {
-      ...answers,
-      sequelizeDialect: dialect ? dialect.name : '',
-    });
-    copyTpl(src('package'), dest('package.json'), {
-      ...answers,
-      description: JSON.stringify(answers.description),
-    });
-    copyTpl(src('env'), dest('.env'), { ...answers, name: envName });
-    copyTpl(src('env.example'), dest('.env.example'), { ...answers, name: envName });
-    copyTpl(src('src/index'), dest('src/index.js'), answers);
-    copyTpl(src('src/config/index'), dest('src/config/index.js'), { ...answers, name: envName });
-    copyTpl(src('src/api/index'), dest('src/api/index.js'), answers);
+    //
+    // ─── SEQUELIZE ───────────────────────────────────────────────────
+    //
 
     if (answers.sequelize) {
-      copy(src('.sequelizerc'), dest('.sequelizerc'));
-      copy(src('src/db/models/index'), dest('src/db/models/index.js'));
-      copy(src('src/db/migrations'), dest('src/db/migrations'));
-      copy(src('src/db/seeders'), dest('src/db/seeders'));
-      copy(src('src/config/sequelize'), dest('src/config/sequelize.js'));
-      copyTpl(src('src/config/database'), dest('src/config/database.js'), {
-        ...answers,
-        dialect: dialect.value,
-      });
+      this.packages.dependencies.push(
+        'bcryptjs@2',
+        'sequelize@5',
+        'sequelize-cli@5',
+        'sequelize-typescript@1',
+        ...dialect.packages
+      );
+      this.packages.devDependencies.push('@types/bcryptjs@2', '@types/bluebird@3');
 
-      this.packages.dependencies.push('sequelize@5');
-      this.packages.dependencies.push('sequelize-cli@5');
-      this.packages.dependencies.push(...dialect.packages);
+      copy('db/migrations/1-init.js.ejs');
+      copy('db/seeders/empty');
+      copy('db/database.js.ejs');
+      copy('src/models/index.ts.ejs');
+      copy('src/models/user.ts.ejs');
+      copy('src/loaders/sequelize.ts.ejs');
+      copy('.sequelizerc.ejs');
     }
+
+    //
+    // ─── REDIS ───────────────────────────────────────────────────────
+    //
+
+    if (answers.redis) {
+      this.packages.dependencies.push('redis@3');
+      this.packages.devDependencies.push('@types/redis@2');
+
+      copy('src/loaders/redis.ts.ejs');
+    }
+
+    //
+    // ─── SOCKETIO ────────────────────────────────────────────────────
+    //
+
+    if (answers.socketIo) {
+      this.packages.dependencies.push(
+        '@ssnxd/socketio-jwt@4',
+        'socket.io@2',
+        'socket.io-client@2.3'
+      );
+      this.packages.devDependencies.push('@types/socket.io@2', '@types/socket.io-client@1');
+
+      copy('src/loaders/socket-io.ts.ejs');
+    }
+
+    //
+    // ─── I18N ────────────────────────────────────────────────────────
+    //
+
+    if (answers.i18n) {
+      this.packages.dependencies.push('i18next@19', 'i18next-http-middleware@3');
+
+      copy('src/locale/en.json.ejs');
+      copy('src/locale/fr.json.ejs');
+      copy('src/locale/index.ts.ejs');
+      copy('src/loaders/i18n.ts.ejs');
+    }
+
+    //
+    // ─── WINSTON ─────────────────────────────────────────────────────
+    //
 
     if (answers.winston) {
-      copy(src('src/config/logger'), dest('src/config/logger.js'));
-
       this.packages.dependencies.push('winston@3');
+
+      copy('src/config/logger.ts.ejs');
     }
+
+    //
+    // ─── CELEBRATE ───────────────────────────────────────────────────
+    //
 
     if (answers.celebrate) {
-      copy(src('src/api/middlewares/validation'), dest('src/api/middlewares/validation.js'));
+      this.packages.dependencies.push('@hapi/joi@15', 'celebrate@10');
+      this.packages.devDependencies.push('@types/hapi__joi@15');
 
-      this.packages.dependencies.push('celebrate@10');
+      copy('src/middlewares/validation.ts.ejs');
     }
+
+    //
+    // ─── JWT ─────────────────────────────────────────────────────────
+    //
 
     if (answers.jwt) {
-      this.packages.dependencies.push('jsonwebtoken@8.5');
-      this.packages.dependencies.push('express-jwt@5.3');
-      this.packages.dependencies.push('uuid@7.0');
+      this.packages.dependencies.push('jsonwebtoken@8', 'express-jwt@6', 'uuid@8');
+      this.packages.devDependencies.push(
+        '@types/express-jwt@0',
+        '@types/jsonwebtoken@8',
+        '@types/uuid@8'
+      );
 
-      copy(src('src/utils/jwt'), dest('src/utils/jwt.js'));
-      copy(src('src/api/middlewares/auth'), dest('src/api/middlewares/auth.js'));
+      copy('src/utils/jwt.ts.ejs');
+      copy('src/middlewares/jwt.ts.ejs');
+      copy('src/types/jwt.ts.ejs');
     }
+
+    //
+    // ─── NODEMAILER ──────────────────────────────────────────────────
+    //
 
     if (answers.nodemailer) {
       this.packages.dependencies.push('nodemailer@6');
+      this.packages.devDependencies.push('@types/nodemailer@6');
 
-      copy(src('src/config/smtp'), dest('src/config/smtp.js'));
-      copyTpl(src('src/utils/mail'), dest('src/utils/mail.js'), answers);
+      copy('src/loaders/mailer.ts.ejs');
+      copy('src/services/email.ts.ejs');
     }
+
+    //
+    // ─── CRON ────────────────────────────────────────────────────────
+    //
+
+    if (answers.cron) {
+      this.packages.dependencies.push('node-schedule@1', '@types/node-schedule@1');
+
+      copy('src/jobs/index.ts.ejs');
+      copy('src/loaders/job-scheduler.ts.ejs');
+    }
+
+    //
+    // ─── ESLINT ──────────────────────────────────────────────────────
+    //
+
+    if (answers.eslint) {
+      this.packages.devDependencies.push(
+        '@typescript-eslint/eslint-plugin@4',
+        '@typescript-eslint/parser@4',
+        'eslint@7',
+        'eslint-config-airbnb-typescript@11',
+        'eslint-import-resolver-typescript@2',
+        'eslint-plugin-import@2',
+        'eslint-plugin-node@8',
+        'eslint-plugin-promise@4'
+      );
+
+      copy('.eslintrc.json.ejs');
+      copy('.eslintignore.ejs');
+    }
+
+    //
+    // ─── PRETTIER ────────────────────────────────────────────────────
+    //
 
     if (answers.prettier) {
-      copy(src('prettierrc'), dest('.prettierrc'));
+      this.packages.devDependencies.push(
+        'eslint-config-prettier@6',
+        'eslint-plugin-prettier@3',
+        'prettier@2'
+      );
 
-      this.packages.devDependencies.push('eslint-config-prettier@latest');
-      this.packages.devDependencies.push('eslint-plugin-prettier@latest');
-      this.packages.devDependencies.push('prettier@latest');
+      copy('.prettierrc.ejs');
+      copy('.prettierignore.ejs');
     }
+
+    //
+    // ─── MOCHA ───────────────────────────────────────────────────────
+    //
 
     if (answers.mocha) {
-      copy(src('mocharc'), dest('.mocharc.json'));
-      copy(src('nycrc'), dest('.nycrc'));
-      copy(src('test/index'), dest('test/index.js'));
-      copy(src('test/api/index'), dest('test/api/index.js'));
-      copyTpl(src('env.test'), dest('.env.test'), { ...answers, name: envName });
+      this.packages.devDependencies.push(
+        '@types/sinon@9',
+        '@types/chai@4',
+        '@types/mocha@8',
+        '@types/supertest@2',
+        'chai@4',
+        'mocha@7',
+        'nyc@15',
+        'sinon@9',
+        'supertest@4',
+        'eslint-plugin-chai-expect@2'
+      );
 
-      this.packages.devDependencies.push('chai@4');
-      this.packages.devDependencies.push('mocha@7');
-      this.packages.devDependencies.push('nyc@15');
-      this.packages.devDependencies.push('supertest@4');
-      this.packages.devDependencies.push('eslint-plugin-chai-expect@2');
+      copy('.mocharc.json.ejs');
+      copy('.nycrc.ejs');
+      copy('.env.test.ejs');
+      copy('test/helpers/stubs/index.ts.ejs');
+      copy('test/helpers/http-responses.ts.ejs');
+      copy('test/helpers/index.ts.ejs');
+      copy('test/integration/404.test.ts.ejs');
+      copy('test/integration/health.test.ts.ejs');
+      copy('test/setup.ts.ejs');
+
+      if (answers.sequelize) {
+        copy('test/helpers/truncate.ts.ejs');
+      }
+
+      if (answers.redis) {
+        this.packages.devDependencies.push('redis-mock@0', '@types/redis-mock@0');
+
+        copy('test/helpers/stubs/redis.ts.ejs');
+      }
+
+      if (answers.nodemailer) {
+        copy('test/helpers/stubs/mailer.ts.ejs');
+        copy('test/unit/mailer.test.ts.ejs');
+      }
+
+      if (answers.sentry) {
+        copy('test/integration/sentry.test.ts.ejs');
+      }
+
+      if (answers.docker) {
+        copy('docker-compose.test.yml.ejs');
+      }
+
+      if (answers.auth) {
+        copy('test/integration/api/auth/post.login.test.ts.ejs');
+      }
     }
+
+    //
+    // ─── SENTRY ──────────────────────────────────────────────────────
+    //
+
+    if (answers.sentry) {
+      this.packages.dependencies.push('@sentry/node@5', '@sentry/tracing@5');
+
+      copy('src/config/sentry.ts.ejs');
+
+      if (answers.winston) {
+        this.packages.dependencies.push('winston-transport-sentry-node@0.7');
+      }
+    }
+
+    //
+    // ─── DOCKER ──────────────────────────────────────────────────────
+    //
 
     if (answers.docker) {
-      copy(src('.dockerignore'), dest('.dockerignore'));
-      copyTpl(src('DockerFile'), dest('DockerFile'), answers);
+      copy('docker/nginx/nginx.tmpl.ejs');
+      copy('docker/node/Dockerfile.ejs');
+      copy('docker/start.sh.ejs');
+      copy('docker-compose.yml.ejs');
+      copy('docker-compose.dev.yml.ejs');
+      copy('docker-compose.prod.yml.ejs');
+      copy('.dockerignore.ejs');
+
+      if (answers.sequelize) {
+        copy('docker/db/Dockerfile.ejs');
+        copy('docker/db/my.cnf.ejs');
+      }
     }
+
+    //
+    // ─── OPENAPI ─────────────────────────────────────────────────────
+    //
 
     if (answers.openapi) {
-      copyTpl(src('src/api/doc/index.html'), dest('src/api/doc/index.html'), answers);
-      copyTpl(src('src/api/doc/openapi'), dest('src/api/doc/openapi.yaml'), answers);
+      copy('src/public/doc/index.html.ejs');
+      copy('src/public/doc/openapi.yml.ejs');
     }
 
-    // Check for empty directories not copied
-    if (!(answers.jwt || answers.celebrate)) {
-      copy(src('src/api/middlewares/empty'), dest('src/api/middlewares/empty'));
+    //
+    // ─── AUTH ────────────────────────────────────────────────────────
+    //
+
+    if (answers.auth) {
+      copy('src/routes/auth.ts.ejs');
+      copy('src/services/auth.ts.ejs');
+      copy('src/types/auth.ts.ejs');
     }
 
-    if (!(answers.jwt || answers.nodemailer)) {
-      copy(src('src/utils/empty'), dest('src/utils/empty'));
+    //
+    // ─── CHECK FOR EMPTY DIRECTORIES NOT COPIED ──────────────────────
+    //
+
+    if (!answers.jwt) {
+      copy('src/utils/empty');
+    }
+
+    if (answers.mocha && !answers.nodemailer) {
+      copy('test/unit/empty');
+    }
+
+    if (answers.mocha && !answers.auth) {
+      copy('test/integration/api/empty');
     }
   }
 
   install() {
-    const appDir = join(process.cwd(), this.answers.shortname);
+    const appDir = path.join(process.cwd(), this.answers.shortname);
 
     process.chdir(appDir);
 
     this.installDependencies({ bower: false, npm: true }).then(() => {
       this.spawnCommandSync('npm', ['i', '--save', ...this.packages.dependencies]);
       this.spawnCommandSync('npm', ['i', '--save-dev', ...this.packages.devDependencies]);
-      this.spawnCommandSync('npm', ['run', 'lint:fix']);
+      this.spawnCommandSync('npm', [
+        'run',
+        this.answers.eslint || this.answers.prettier ? 'lint:fix' : 'lint',
+      ]);
     });
   }