+++ generators/app/utils/sequelize-dialect-enum.js
@@ -1,27 +1,26 @@
 module.exports = {
-  MYSQL: {
+  mysql: {
     name: 'MySQL',
     value: 'mysql',
     packages: ['mysql2@2'],
+    port: 3306,
   },
-  POSTGRESQL: {
+  postgresql: {
     name: 'PostgreSQL',
     value: 'postgres',
     packages: ['pg@7', 'pg-hstore'],
+    port: 5432,
   },
-  SQLITE: {
-    name: 'SQLite',
-    value: 'sqlite',
-    packages: ['sqlite3@4'],
-  },
-  MARIA: {
+  maria: {
     name: 'MariaDB',
     value: 'mariadb',
     packages: ['mariadb'],
+    port: 3306,
   },
-  MSSQL: {
+  mssql: {
     name: 'Microsoft SQL Server',
     value: 'mssql',
     packages: ['tedious@6'],
+    port: 1433,
   },
 };